Update README.md
This assignment asks you to examine the relationship between median house values and several neighborhood characteristics, using Philadelphia data at the Census block group level. You will need R for this assignment. Remember that this report needs to be written with an introduction, methods/results, and discussion. Pointers for a successful report, as well as an example on how to interpret regression output, are included. Also included is an outline which you’re asked to follow when writing your report. Your submission may be written in MS Word (submitted as a .docx or .pdf file) or in R Markdown format.

Data Description

The 2000 Philadelphia Census block group level dataset RegressionData.csv contains, among other variables, the variables below:

    POLY_ID: Census Block Group ID

    MEDHVAL: Median value of all owner occupied housing units

    PCBACHMORE: Proportion of residents in Block Group with at least a bachelor’s degree

    PCTVACANT: Proportion of housing units that are vacant

    PCTSINGLES: Percent of housing units that are detached single family houses

    NBELPOV100: Number of households with incomes below 100% poverty level (i.e., number of households living in poverty)

    MEDHHINC: Median household income